requirements.txt
facenet-pytorch
Pillow
ttkthemes
customtkinter
tkvideoplayer
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
